requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
attrs==24.3.0
binaryornot==0.4.4
blinker==1.9.0
bracex==2.5.post1
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filetype==1.2.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latexcodec==3.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monty==2025.1.3
mpmath==1.3.0
mrcfile==1.5.3
narwhals==1.21.1
networkx==3.4.2
numpy==2.2.1
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybtex==0.24.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pymatgen==2024.11.13
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.0
six==1.17.0
smmap==5.0.2
spglib==2.5.0
st-circular-progress==0.3.9
streamlit==1.41.1
streamlit-ace==0.1.1
streamlit-antd==0.8.3
streamlit-embeded==0.0.1
streamlit-file-browser==3.2.22
streamlit-molstar==0.4.21
streamlit-pdf-viewer==0.0.20
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uncertainties==3.2.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
wcmatch==10.0
